requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
sentence-transformers
numpy
python-dotenv
requests
requests
